# api-services/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Sell It App",
    "description": "This is going to make it easy for the front end guys to play around with the AP",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api/v1/",
      "description": "Local Dev"
    }
  ],
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/chat/{matrikel_number}": {
      "parameters": [
        {
          "name": "matrikel_number",
          "in": "path",
          "required": true,
          "description": "Get all the chat for a particular matrikel_number",
          "type": "number"
        }
      ],
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "Get all the chat for a particular matrikel_number",
        "parameters": [
          {
            "in": "path",
            "name": "matrikel_number",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/matrikel_number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Chat"
            }
          }
        }
      }
    },
    "/chat/sendMessage": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Send a Message",
        "requestBody": {
          "description": "SendChat Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/SendChat"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/id"
            }
          },
          "400": {
            "description": "Failed. Bad post data."
          }
        }
      }
    },
    "/chat/getAllMessages": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Get All Message",
        "requestBody": {
          "description": "GetChat Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/GetChat"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/id"
            }
          },
          "400": {
            "description": "Failed. Bad post data."
          }
        }
      }
    }
  },
  "definitions": {
    "id": {
      "properties": {
        "id": {
          "type": "number"
        }
      }
    },
...
    "Chat": {
      "type": "object",
      "properties": {
        "matrikel_number": {
          "type": "number"
        },
        "first_name": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "product_name": {
          "type": "string"
        }
      }
    },
    "SendChat": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "product_id": {
          "type": "number"
        },
        "sender_id": {
          "type": "number"
        },
        "receiver_id": {
          "type": "number"
        }
      }
    },
    "GetChat": {
      "type": "object",
      "properties": {
        "receiver_id": {
          "type": "number"
        },
        "sender_id": {
          "type": "number"
        }
      }
    }
  }
}